pyproject.toml
###############################################################################
# Caterva - On demand access to remote Blosc2 data repositories
#
# Copyright (c) 2023 The Blosc Developers <blosc@blosc.org>
# https://www.blosc.org
# License: GNU Affero General Public License v3.0
# See LICENSE.txt for details about copyright and rights to use.
###############################################################################

[project]
name = "Caterva"
requires-python = ">3.8"
version = "0.1.0"
readme = "README.md"
authors = [
    {name = "Blosc Development Team", email = "contact@blosc.org"},
]
description = "Caterva - On demand access to remote Blosc2 data repositories"
keywords = ["pubsub", "blosc2"]
license = {text = "GNU Affero General Public License version 3"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
]
dependencies = [
    "blosc2",
    "fastapi_websocket_pubsub",
    "httpx",
    "uvicorn",
    "watchfiles",
]

[tool.setuptools]
platforms = [ "any" ]

[tool.scikit-build]
wheel.py-api = "py3"

[tool.pytest.ini_options]
testpaths = [
    "tests/test.py",
]
